Telegram_chatbot/requirements.txt
aiogram>=3.22.0
python-dotenv>=1.0.0
openai>=0.28.0
